Add Tags
Add tags to my personal blog. This commit adds the tags to the individual
blogs as well as creates two new pages. A /tags route which lists
all of the tags and the counts of those tags, and it also adds
a tag display page which will display all the blog posts with an
associated tag. This commit has done some refactoring also to
make things slightly more modular since we now use the same components
throughout multiple pages.